Moving typing to types
