+++ src/main/scala/wdlTools/eval/Eval.scala
@@ -5,7 +5,7 @@ import java.net.URL
 import wdlTools.eval.WdlValues._
 import wdlTools.syntax.{AbstractSyntax => AST}
 import wdlTools.syntax.{TextSource, WdlVersion}
-import wdlTools.typing.WdlTypes
+import wdlTools.types.WdlTypes
 import wdlTools.util.{EvalConfig, Options}
 
 case class Eval(opts: Options,
