+++ src/main/scala/wdlTools/eval/StdlibV1.scala
@@ -5,7 +5,7 @@ import java.nio.file.{Path, Paths}
 import wdlTools.eval.WdlValues._
 import wdlTools.syntax.TextSource
 import wdlTools.util.{EvalConfig, Options}
-import wdlTools.typing.WdlTypes._
+import wdlTools.types.WdlTypes._
 
 case class StdlibV1(opts: Options, evalCfg: EvalConfig, docSourceURL: Option[URL])
     extends StandardLibraryImpl {
@@ -276,7 +276,16 @@ case class StdlibV1(opts: Options, evalCfg: EvalConfig, docSourceURL: Option[URL
   // File write_object(Object)
   //
   private def write_object(args: Vector[WV], text: TextSource): WV_File = {
-    throw new EvalException("write_object: not implemented", text, docSourceURL)
+    assert(args.size == 1)
+    val obj = coercion.coerceTo(WT_Object, args.head, text).asInstanceOf[WV_Object]
+    val keyLine = obj.members.keys.mkString("\t")
+    val valueLine = obj.members.values.map{ vw =>
+      coercion.coerceTo(WT_String, vw, text).asInstanceOf[WV_String]
+    }.mkString("\t")
+    val content = keyLine + "\n" + valueLine
+    val tmpFile: Path = iosp.mkTempFile()
+    iosp.writeFile(tmpFile, content)
+    WV_File(tmpFile.toString)
   }
 
   private def write_objects(args: Vector[WV], text: TextSource): WV_File = {
