+++ src/main/scala/wdlTools/types/WdlTypes.scala
@@ -1,4 +1,4 @@
-package wdlTools.typing
+package wdlTools.types
 
 // This is the WDL typesystem
 object WdlTypes {
