+++ src/main/scala/wdlTools/types/package.scala
@@ -1,4 +1,4 @@
-package wdlTools.typing
+package wdlTools.types
 
 import java.net.URL
 import wdlTools.syntax.TextSource
