+++ src/test/resources/types/v1/coercions_strict.wdl
