+++ src/test/resources/types/v1/command_string.wdl
