+++ src/test/resources/types/v1/compound_expr_bug.wdl
