+++ src/test/resources/types/v1/meta_section_compound.wdl
