+++ src/test/resources/types/v1/scatter_I.wdl
