+++ src/test/resources/types/v1/shadow_II.wdl
