+++ src/test/resources/types/v1/stdlib.wdl
