+++ src/test/scala/wdlTools/typing/package.scala
@@ -1,4 +1,4 @@
-package wdlTools.typing
+package wdlTools.types
 
 import org.scalatest.Tag
 object Edge extends Tag("edge")
